+++ forge/src/main/java/de/cech12/usefulhats/platform/ForgeRegistryHelper.java
@@ -62,6 +62,11 @@ public class ForgeRegistryHelper implements IRegistryHelper {
     }
 
     @Override
+    public boolean isEntityInFluid(LivingEntity entity) {
+        return entity.isInFluidType();
+    }
+
+    @Override
     public boolean isAxe(ItemStack tool) {
         return tool.canPerformAction(ToolActions.AXE_DIG);
     }